Install Guide for Installing CodeSys on a Phoenix Contact Web Panel (WP6xxxx)
This is a step by step guide for installing CodeSys on a Phoenix Web Panel (WP6 xxxx, where xxxx defines screen size).
All of the WP6 series Panels will work with this Panel.
NOTE: CodeMeter is currently not compatible with the WP6 Panels, therefore ALL Licensing must be accomplished using a Dongle.  
            Contact CodeSys Directly to Procure this Dongle.

Please note that the CodeSys Web Visu default uses Port 8000, which is already taken by the WP6 panels.  A second Document attached
describes how to modify the standard CodeSys target to Port 8001, to allow the panel to function properly with CodeSys.